requirements.txt
numpy==2.0.2
pandas==2.2.3
pickle-mixin==1.0.2
scikit-learn==1.6.1
streamlit==1.43.1
tensorflow==2.18.0
opencv-python-headless==4.11.0.86
pillow==11.1.0
matplotlib==3.10.1
